+++ platforms/browser/platform_www/cordova_plugins.js
@@ -152,6 +152,44 @@ module.exports = [
         "runs": true
     },
     {
+        "file": "plugins/cordova-sqlite-storage/www/SQLitePlugin.js",
+        "id": "cordova-sqlite-storage.SQLitePlugin",
+        "pluginId": "cordova-sqlite-storage",
+        "clobbers": [
+            "SQLitePlugin"
+        ]
+    },
+    {
+        "file": "plugins/cordova-plugin-splashscreen/src/browser/SplashScreenProxy.js",
+        "id": "cordova-plugin-splashscreen.SplashScreenProxy",
+        "pluginId": "cordova-plugin-splashscreen",
+        "runs": true
+    },
+    {
+        "file": "plugins/cordova-plugin-splashscreen/www/splashscreen.js",
+        "id": "cordova-plugin-splashscreen.SplashScreen",
+        "pluginId": "cordova-plugin-splashscreen",
+        "clobbers": [
+            "navigator.splashscreen"
+        ]
+    },
+    {
+        "file": "plugins/cordova-plugin-appversion/www/app-version.js",
+        "id": "cordova-plugin-appversion.RareloopAppVersion",
+        "pluginId": "cordova-plugin-appversion",
+        "clobbers": [
+            "AppVersion"
+        ]
+    },
+    {
+        "file": "plugins/cordova-plugin-app-update/www/AppUpdate.js",
+        "id": "cordova-plugin-app-update.AppUpdate",
+        "pluginId": "cordova-plugin-app-update",
+        "clobbers": [
+            "AppUpdate"
+        ]
+    },
+    {
         "file": "plugins/cordova-plugin-googlemaps/www/Promise.js",
         "id": "cordova-plugin-googlemaps.Promise",
         "pluginId": "cordova-plugin-googlemaps"
@@ -273,6 +311,24 @@ module.exports = [
     {
         "file": "plugins/cordova-plugin-googlemaps/www/LocationService.js",
         "id": "cordova-plugin-googlemaps.LocationService",
+        "pluginId": "cordova-plugin-googlemaps",
+        "runs": true
+    },
+    {
+        "file": "plugins/cordova-plugin-googlemaps/www/ElevationService.js",
+        "id": "cordova-plugin-googlemaps.ElevationService",
+        "pluginId": "cordova-plugin-googlemaps",
+        "runs": true
+    },
+    {
+        "file": "plugins/cordova-plugin-googlemaps/www/DirectionsService.js",
+        "id": "cordova-plugin-googlemaps.DirectionsService",
+        "pluginId": "cordova-plugin-googlemaps",
+        "runs": true
+    },
+    {
+        "file": "plugins/cordova-plugin-googlemaps/www/DirectionsRenderer.js",
+        "id": "cordova-plugin-googlemaps.DirectionsRenderer",
         "pluginId": "cordova-plugin-googlemaps",
         "runs": true
     },
@@ -453,6 +509,18 @@ module.exports = [
         "runs": true
     },
     {
+        "file": "plugins/cordova-plugin-googlemaps/src/browser/PluginElevationService.js",
+        "id": "cordova-plugin-googlemaps.PluginElevationService",
+        "pluginId": "cordova-plugin-googlemaps",
+        "runs": true
+    },
+    {
+        "file": "plugins/cordova-plugin-googlemaps/src/browser/PluginDirectionsService.js",
+        "id": "cordova-plugin-googlemaps.PluginDirectionsService",
+        "pluginId": "cordova-plugin-googlemaps",
+        "runs": true
+    },
+    {
         "file": "plugins/cordova-plugin-googlemaps/src/browser/PluginGeocoder.js",
         "id": "cordova-plugin-googlemaps.PluginGeocoder",
         "pluginId": "cordova-plugin-googlemaps",
@@ -463,44 +531,6 @@ module.exports = [
         "id": "cordova-plugin-googlemaps.PluginEnvironment",
         "pluginId": "cordova-plugin-googlemaps",
         "runs": true
-    },
-    {
-        "file": "plugins/cordova-sqlite-storage/www/SQLitePlugin.js",
-        "id": "cordova-sqlite-storage.SQLitePlugin",
-        "pluginId": "cordova-sqlite-storage",
-        "clobbers": [
-            "SQLitePlugin"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-splashscreen/src/browser/SplashScreenProxy.js",
-        "id": "cordova-plugin-splashscreen.SplashScreenProxy",
-        "pluginId": "cordova-plugin-splashscreen",
-        "runs": true
-    },
-    {
-        "file": "plugins/cordova-plugin-splashscreen/www/splashscreen.js",
-        "id": "cordova-plugin-splashscreen.SplashScreen",
-        "pluginId": "cordova-plugin-splashscreen",
-        "clobbers": [
-            "navigator.splashscreen"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-appversion/www/app-version.js",
-        "id": "cordova-plugin-appversion.RareloopAppVersion",
-        "pluginId": "cordova-plugin-appversion",
-        "clobbers": [
-            "AppVersion"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-app-update/www/AppUpdate.js",
-        "id": "cordova-plugin-app-update.AppUpdate",
-        "pluginId": "cordova-plugin-app-update",
-        "clobbers": [
-            "AppUpdate"
-        ]
     }
 ];
 module.exports.metadata = 
@@ -523,13 +553,13 @@ module.exports.metadata =
     "cordova-plugin-androidx": "1.0.2",
     "cordova.plugins.diagnostic": "5.0.2",
     "phonegap-plugin-barcodescanner": "8.1.0",
-    "com.googlemaps.ios": "3.8.0",
-    "cordova-plugin-googlemaps": "2.7.1",
     "cordova-plugin-androidx-adapter": "1.1.1",
-    "cordova-sqlite-storage": "5.0.0",
+    "cordova-sqlite-storage": "5.0.1",
     "cordova-plugin-splashscreen": "6.0.0",
     "cordova-plugin-appversion": "1.0.0",
-    "cordova-plugin-app-update": "2.0.2"
+    "cordova-plugin-app-update": "2.0.2",
+    "com.googlemaps.ios": "3.9.0",
+    "cordova-plugin-googlemaps": "2.8.0-20200709-2008"
 }
 // BOTTOM OF METADATA
 });