+++ plugins/cordova-plugin-googlemaps/src/android/plugin/google/maps/MyPlugin.java
@@ -28,6 +28,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+
 public class MyPlugin extends CordovaPlugin implements MyPluginInterface {
   public MyPlugin self = null;
   public final Map<String, Method> methods = new ConcurrentHashMap<String, Method>();