+++ plugins/cordova-plugin-googlemaps/src/android/plugin/google/maps/PluginLocationService.java
@@ -36,7 +36,6 @@ import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
-
 public class PluginLocationService extends CordovaPlugin {
   private Activity activity;
   private final String TAG = "PluginLocationService";