+++ src/app/interfaces/meta.ts
@@ -1,0 +1,5 @@
+export interface Meta {
+    label: string,
+    lt?: number,
+    lg?: number
+}